refactor : error 로그 찍기
